Services ready, remaining endpoints

+++ backend/src/main/java/com/example/backend/airline_stat/strategy/StatisticsStrategy.java
@@ -2,5 +2,5 @@ package com.example.backend.airline_stat.strategy;
 
 
 public interface StatisticsStrategy {
-    Long calculate(String airlineName);
+    Double calculate(String airlineName);
 }
